Add files via upload
The aim is to find a tour yielding the maximum path cost. In this problem, we try to solve Santa’s problem using two methods. The first method is finding a path that yields minimum path cost, the other method is supposed to provide two disjoint paths instead of one. 
The genetic algorithm gave the best results to obtain the shortest path.